fix provider reopener, rework nonce cache (#481)
The provider factory reopener's concept of "last_consistent_block" was
incorrect, as database consistency was compromised for previously
checked blocks. This caused the rbuilder to miss some slots due to its
inability to reopen the database during slot execution.

The nonce cache has been refactored to use database transactions instead
of database handles. This change simplifies the code and reduces the
need for database consistency checks.

## 📝 Summary

<!--- A general summary of your changes -->

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)

---------

Co-authored-by: Daniel Xifra <mandinga@gmail.com>