fix coinbase payment edge case (#478)
## 📝 Summary

When the coinbase address is the fee recipient, we exclusively use
fee_recipient_balance_diff
since this is the value used by validation nodes

Using fee_recipient_balance_diff may cause block validation failures in
certain edge cases

Real Example: If the fee recipient is a contract that sweeps its balance
to another address on each call,
and we include a bundle paying directly to coinbase, the fee recipient
balance would be 0 causing validation nodes to reject the block

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)