Order sorting using types. (#493)
## 📝 Summary

This is a refactoring PR, completely backwards compatible.
Order sorting was based on the ability to express the sorting as a
number (Sorting::sorting_value) and this is way too rigid.
Sorting::sorting_value was replaced for type system (implementing
OrderPriority). Now we can create new sorting criteria by adding new
types.
I also added a patch to remove the internally hardcoded check for
"enforce_inplace_sim_result". Now it's explicit and configurable (we'll
need it in the future)

## 💡 Motivation and Context

This allows to easily add new sorting criteria and more advanced
handling for the "enforce_inplace_sim_result".

## ✅ I have completed the following steps:

* [X] Run `make lint`
* [X] Run `make test`
* [ ] Added tests (if applicable)