Fixes a regression that made orderings non-deterministic (#515)
## 📝 Summary

Before new ordering system was introduced orderings were always
deterministic because they were breaking ties using order ids and now we
don't do that and that PR returns that functionality. Without this
back-tests and refunds are non-deterministic which is bad for
development but also makes refund less precise because live builder also
does not have this level of determinism.

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)