fix: don't miss blocks on batcher updates (#529)
## 📝 Summary

When batcher's max channel duration is big enough (e.g. 10m), the
batcher would be pushing its updates at specified intervals.  This
causes the sequencer to send an avalanche of FCUs (and
getBlockByNumber) that push safe head step-by-step.  As a consequence
it can happen that the time b/w FCU and ensuing getPayload would be
on the scale of ~2.5s.  This means that we should "remember" the
payloads long enough to accommodate that corner-case.

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)