Feat: `evm` module with custom `EvmFactory` trait (#554)
## 💡 Motivation and Context

This PR brings small code cleanup for `EvmFactory`. I have created a new
`EvmFactory` trait in rbuilder because the `Evm` trait from `alloy_revm`
has a lot of associated types and trait bounds.
These are now defined in the custom `EvmFactory` trait and resolved in
the implementation for `EthCachedEvmFactory`, so this removes the need
for extra trait bounds outside of the `evm` module (such as in the
`execute_evm` function in `order_commit`).

## 📝 Summary

- The `EthCachedEvmFactory` was moved to new module `evm` under
`building` (`rbuilder::building::evm`), IMO this makes the overall code
structure better as the separation of concerns is cleaner (more obvious)
- `EthCachedEvmFactory::Context` is defined to `EthEvmContext<DB>`,
which is an alias of `Context<BlockEnv, TxEnv, CfgEnv, DB>`
- `EvmFactory` custom trait, which is a more concrete version than
`alloy_revm::EvmFactory`
- Remove the dependency on `EthEvmFactory` in `EthCachedEvmFactory`
struct, but still use it's default implementation under the hood (we can
also consider completely remove it and use `Context` instead)

---

## ✅ I have completed the following steps:

* [x] Run `make lint`
* [x] Run `make test`
* [ ] Added tests (if applicable)