Cache bloom filters (#598)
## 📝 Summary

Currently finalize time is often 30% bigger than root hash time but
ideally it should be really close.

One of the reasons is that we calculate bloom filters every single
finalize which can be expensive operation, this PR introduces caching
for bloom filters.

It also moves finalize to a separate thread so it can persist local
thread cache for the slot finalization and in addition the thread is not
named which makes profiling easier.

This shaves up to 5ms in the worst case from finalize but its a good
start for faster finalization

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)