Refactoring unifying tx+receipt+gas into a struct TransactionExecutionInfo. (#620)
## 📝 Summary

In a few places we had a vector of tx and a brother vector of receipt
matching one the txs.
I put both in a new struct TransactionExecutionInfo and I also moved
there gas and added profit.
Some copy paste was removed.

## 💡 Motivation and Context

Nicer code and it will allow better checks on bundle execution due to
the new per tx profit info.

## ✅ I have completed the following steps:

* [X] Run `make lint`
* [X] Run `make test`
* [ ] Added tests (if applicable)