Fusaka (#698)
## 📝 Summary

Changes for Fusaka fork:
- Had to upgrade to rust 1.88: lint hell.
- alloy/nybbles: https://github.com/alloy-rs/nybbles/pull/17 updates the
internal representation of Nibble from a SmallVec<[u8; 64]> to a U256.
This breaks code in the merkle eth sparse tree as the code indexes a lot
into the SmallVec representation.
For this version we kept the sparse tree internals and adapted the on
the interface (we have deps in both nibble versions).
- New Eip7594 blobs: This blobs have more profs so now we use
BlobTransactionSidecarVariant instead of BlobTransactionSidecar.
BlobTransactionSidecarVariant can model the old blob
(BlobTransactionSidecar) and the new one
(BlobTransactionSidecarEip7594).
- revm changed BlockNumber and Timestamp from U256 instead of u64. We
convert internally to 64 since it's more compatible with the rest.
- Minor traits changes on reth.
- New relay SignedBidSubmissionV5 with new BlobsBundleV2.
- New rlp block encoding limiting as specified by EIP-7934.
- Order filter to avoid mix of blob types before and after the fork.

Thanks @bharath-123 for the help!

## ✅ I have completed the following steps:

* [X] Run `make lint`
* [X] Run `make test`
* [ ] Added tests (if applicable)