fix: preload non-existent proof targets (#706)
## 📝 Summary

Previously, generating proofs for accounts missing from bundle state did
not work, because the state root algorithm requires account info to be
present and does not have access to state provider to be able to fetch.
Impose a requirement on the caller to pre-load proof target accounts
into bundle state. Document the change in the behavior.

## ✅ I have completed the following steps:

* [x] Run `make lint`
* [x] Run `make test`
* [ ] Added tests (if applicable)