fix: don't filter out regular txs post fusaka (#709)
## 📝 Summary

The blob filter activated post fusaka was filtering out non-eip4844 txs.
This is because it only keeps transactions with
`BlobTransactionSidecarVariant::Eip7594(_)`. But non-eip4844 txs have
`BlobTransactionSidecarVariant::Eip4844(_)` by default with 0 blobs,
commitments and proofs.

Because of this, post Fusaka only blob txs were being packed in the
block since other txs were being rejected by this condiction.

To fix this, we pass the entire tx to the filter function and only check
what type of `BlobTransactionSidecarVariant` if the tx is an eip4844 tx.

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)