Better bidding 2 (#711)
## 📝 Summary

Adds a new way to seal blocks to enable it set `adjust_finalized_blocks
= true` in the config.

If set to true blocks will be processed in two steps:
1. block is finalized (root hash, receipts root, etc) is processed
2. finalized blocks is sent to the bidding service
3. When bidding service responds finalized block is adjusted with
correct bid value

The advantage is that bidding response becomes faster (it takes around
0.5ms on average to adjust finalized block).

In addition this PR:
* adds blob gas to the BlockSpace structure
* allows only transactions payment as a way to pay a proposer 

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)