feat(primitives): decoding with signer lookups (#760)
## 📝 Summary

Modifies the `rbuilder-primitives` crate so that it is possible to
decode bundles specifying a signer lookup, to save on signature recovery
compute.

## 💡 Motivation and Context

Most bundles received contain transactions already heard, as such we can
reduce a lot of compute from the hot path.

**Summary of the changes**:
- `RawBundle` now feature a `decode_with_signer_lookup` that allows to
specify a closure to check whether the signer is already known for a
certain hash.
- It calls an underlying `RawBundle::decode_inner` that takes this the
closure as an `Option`.
- Inside `RawBundle::decode_inner`, an helper struct
`RawTransactionDecodable` (naming can be improved) that contains the raw
bytes, the `TxEncoding` and the optional signer lookup is used for
decoding. Transaction decoding and recovering can't really be decoupled
from bundle decoding because it's required for parsing refunds data.
- A `RawBundleMetadata` struct has been introduced to simplify a bit the
code and for separation of concerns.

---

## ✅ I have completed the following steps:

* [x] Run `make lint`
* [x] Run `make test`
* [ ] Added tests (if applicable)