Add titan relay to the relay list (#679)
## 📝 Summary

Add Titan Relay to the relay list

## 💡 Motivation and Context

I added the Titan Relay to the current relay list because it was not
included.
Although the request to the Eden relay is redirected to Titan, I found
that they are not exactly the same in some cases. For example, I
received an empty response from
https://relay.edennetwork.io/relay/v1/data/bidtraces/proposer_payload_delivered?slot=12359595
but a valid response from
https://titanrelay.xyz/relay/v1/data/bidtraces/proposer_payload_delivered?slot=12359595.

---

## ✅ I have completed the following steps:

* [x] Run `make lint`
* [x] Run `make test`
* [ ] Added tests (if applicable)